+++ app/bms/sales/quotations/print_quotation.php
@@ -553,10 +553,12 @@ if ($cust_address !== '') {
                 <span>-<?= $currency ?> <?= number_format($order['discount_amount'], 2) ?></span>
             </div>
             <?php endif; ?>
+            <?php if (floatval($order['tax_amount']) > 0): ?>
             <div class="totals-row">
-                <span>VAT (18%):</span>
+                <span>VAT:</span>
                 <span><?= $currency ?> <?= number_format($order['tax_amount'], 2) ?></span>
             </div>
+            <?php endif; ?>
             <?php if (floatval($order['shipping_cost']) > 0): ?>
             <div class="totals-row">
                 <span>Shipping:</span>